Replace links to lines on github by permalinks
Those links intend to point to the exact line where the
configuration is set.

The previous configuration the link was weak, and will point to
different place if someone add or remove code in those files.

This approach link to the line in a specific commit so it will not
change *never*. As cons, it can but outdated, but the user can navigate.